Move and separate scrape dialog
