+++ ui/v2.5/src/components/Shared/ScrapeDialog/ScrapeDialog.tsx
@@ -8,10 +8,9 @@ import {
   FormControl,
   Badge,
 } from "react-bootstrap";
-import { CollapseButton } from "./CollapseButton";
-import { Icon } from "./Icon";
-import { ModalComponent } from "./Modal";
-import isEqual from "lodash-es/isEqual";
+import { CollapseButton } from "../CollapseButton";
+import { Icon } from "../Icon";
+import { ModalComponent } from "../Modal";
 import clone from "lodash-es/clone";
 import { FormattedMessage, useIntl } from "react-intl";
 import {
@@ -21,78 +20,10 @@ import {
   faTimes,
 } from "@fortawesome/free-solid-svg-icons";
 import { getCountryByISO } from "src/utils/country";
-import { CountrySelect } from "./CountrySelect";
-import { StringListInput } from "./StringListInput";
-import { ImageSelector } from "./ImageSelector";
-
-export class ScrapeResult<T> {
-  public newValue?: T;
-  public originalValue?: T;
-  public scraped: boolean = false;
-  public useNewValue: boolean = false;
-
-  public constructor(
-    originalValue?: T | null,
-    newValue?: T | null,
-    useNewValue?: boolean
-  ) {
-    this.originalValue = originalValue ?? undefined;
-    this.newValue = newValue ?? undefined;
-    // NOTE: this means that zero values are treated as null
-    // this is incorrect for numbers and booleans, but correct for strings
-    const hasNewValue = !!this.newValue;
-
-    const valuesEqual = isEqual(originalValue, newValue);
-    this.useNewValue = useNewValue ?? (hasNewValue && !valuesEqual);
-    this.scraped = hasNewValue && !valuesEqual;
-  }
-
-  public setOriginalValue(value?: T) {
-    this.originalValue = value;
-    this.newValue = value;
-  }
-
-  public cloneWithValue(value?: T) {
-    const ret = clone(this);
-
-    ret.newValue = value;
-    ret.useNewValue = !isEqual(ret.newValue, ret.originalValue);
-
-    // #2691 - if we're setting the value, assume it should be treated as
-    // scraped
-    ret.scraped = true;
-
-    return ret;
-  }
-
-  public getNewValue() {
-    if (this.useNewValue) {
-      return this.newValue;
-    }
-  }
-}
-
-// for types where !!value is a valid value (boolean and number)
-export class ZeroableScrapeResult<T> extends ScrapeResult<T> {
-  public constructor(
-    originalValue?: T | null,
-    newValue?: T | null,
-    useNewValue?: boolean
-  ) {
-    super(originalValue, newValue, useNewValue);
-
-    const hasNewValue = this.newValue !== undefined;
-
-    const valuesEqual = isEqual(originalValue, newValue);
-    this.useNewValue = useNewValue ?? (hasNewValue && !valuesEqual);
-    this.scraped = hasNewValue && !valuesEqual;
-  }
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function hasScrapedValues(values: ScrapeResult<any>[]) {
-  return values.some((r) => r.scraped);
-}
+import { CountrySelect } from "../CountrySelect";
+import { StringListInput } from "../StringListInput";
+import { ImageSelector } from "../ImageSelector";
+import { ScrapeResult } from "./scrapeResult";
 
 export interface IHasName {
   name: string | undefined;
